Update Update-ScreenConnectLECert.ps1
Added array element [0] to $importedCert in the event the script runs within 7 days of LE certificate generation and the import attempt results in 2 entries for the same certificate.  By using array element [0], this will only pull the first entry, regardless of the number of entries actually stored in the variable.